requirements.txt
aiogram>=3.7,<4.0
aiosqlite>=0.20
python-dotenv>=1.0
pydantic>=2.8
alembic>=1.13
